+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased]
+### Added
+- `init` model extension option - used to control initialization of property
+### Changed
+- run `set` function on property init
 ## [0.8.38] - 2022-11-28
 ### Fixed
 - Ensure that a change to a calculation triggers default calculations that depend on it